Test action attempt errors

+++ src/Errors/ActionAttempt/TimeoutError.php
@@ -6,7 +6,7 @@ use Seam\Objects\ActionAttempt;
 
 class TimeoutError extends SeamActionAttemptError
 {
-    public function __construct(ActionAttempt $actionAttempt, int $timeout)
+    public function __construct(ActionAttempt $actionAttempt, float $timeout)
     {
         parent::__construct(
             "Timed out waiting for action attempt after {$timeout}s",
